chat model with gemini integration
